E:\github\focusSip

+++ src/FFmpeg/win/x86/include/libavutil/common.h
@@ -1,3 +1,4 @@
+// ABChernic : 修改了某些强制类型转换
 /*
  * copyright (c) 2006 Michael Niedermayer <michaelni@gmx.at>
  *
@@ -159,7 +160,7 @@ static av_always_inline av_const int64_t av_clip64_c(int64_t a, int64_t amin, in
 static av_always_inline av_const uint8_t av_clip_uint8_c(int a)
 {
     if (a&(~0xFF)) return (-a)>>31;
-    else           return a;
+    else           return (uint8_t)a;
 }
 
 /**
@@ -170,7 +171,7 @@ static av_always_inline av_const uint8_t av_clip_uint8_c(int a)
 static av_always_inline av_const int8_t av_clip_int8_c(int a)
 {
     if ((a+0x80U) & ~0xFF) return (a>>31) ^ 0x7F;
-    else                  return a;
+    else                  return (uint8_t)a;
 }
 
 /**
@@ -181,7 +182,7 @@ static av_always_inline av_const int8_t av_clip_int8_c(int a)
 static av_always_inline av_const uint16_t av_clip_uint16_c(int a)
 {
     if (a&(~0xFFFF)) return (-a)>>31;
-    else             return a;
+    else             return (uint16_t)a;
 }
 
 /**
@@ -192,7 +193,7 @@ static av_always_inline av_const uint16_t av_clip_uint16_c(int a)
 static av_always_inline av_const int16_t av_clip_int16_c(int a)
 {
     if ((a+0x8000U) & ~0xFFFF) return (a>>31) ^ 0x7FFF;
-    else                      return a;
+    else                      return (int16_t)a;
 }
 
 /**
